Corrected length type

+++ ZmodDAC1411_Demo_Linux/src/main.cpp
@@ -41,7 +41,7 @@ void dacRampDemo(float offset, float amplitude, float step, uint8_t channel, uin
 	float val;
 	uint32_t valBuf;
 	int16_t valRaw;
-	int length = (int)(amplitude/step) << 2;
+	size_t length = (size_t)(amplitude/step) << 2;
 	int i;
 	if (length > ((1<<14) - 1))
 	{
